Move configuration tests into focused suite
