+++ Tests/QuillCodeAppTests/WorkspaceConfigurationIntegrationTests.swift
@@ -1,0 +1,89 @@
+import XCTest
+import QuillCodeAgent
+import QuillCodeCore
+import QuillCodePersistence
+@testable import QuillCodeApp
+
+@MainActor
+final class WorkspaceConfigurationIntegrationTests: XCTestCase {
+    func testModeAndModelUpdateSelectedThreadAndTopBar() {
+        let thread = ChatThread()
+        let model = QuillCodeWorkspaceModel(root: QuillCodeRootState(
+            threads: [thread],
+            selectedThreadID: thread.id
+        ))
+
+        model.setMode(.review)
+        model.setModel("provider/model")
+
+        XCTAssertEqual(model.selectedThread?.mode, .review)
+        XCTAssertEqual(model.selectedThread?.model, "provider/model")
+        XCTAssertEqual(model.root.topBar.mode, .review)
+        XCTAssertEqual(model.root.topBar.model, "provider/model")
+    }
+
+    func testToggleModelFavoriteUpdatesConfigAndSurface() {
+        let model = QuillCodeWorkspaceModel(root: QuillCodeRootState(
+            config: AppConfig(favoriteModels: ["provider/old"]),
+            topBar: TopBarState(model: TrustedRouterDefaults.synthModel),
+            modelCatalog: TrustedRouterModelCatalog.defaultModels
+        ))
+
+        model.toggleModelFavorite(" z-ai/glm-5.2 ")
+
+        XCTAssertEqual(model.root.config.favoriteModels, ["provider/old", "z-ai/glm-5.2"])
+        XCTAssertEqual(model.surface().topBar.modelCategories.first?.category, "Favorites")
+        XCTAssertEqual(model.surface().topBar.modelCategories.first?.models.map(\.id), ["provider/old", "z-ai/glm-5.2"])
+
+        model.toggleModelFavorite("provider/old")
+
+        XCTAssertEqual(model.root.config.favoriteModels, ["z-ai/glm-5.2"])
+        XCTAssertEqual(model.surface().topBar.modelCategories.first?.models.map(\.id), ["z-ai/glm-5.2"])
+    }
+
+    func testApplySettingsUpdatesConfigThreadAndSettingsSurface() {
+        let thread = ChatThread()
+        let model = QuillCodeWorkspaceModel(root: QuillCodeRootState(
+            threads: [thread],
+            selectedThreadID: thread.id
+        ))
+        let config = AppConfig(
+            defaultModel: "z-ai/glm-5.2",
+            mode: .review,
+            apiBaseURL: "https://api.trustedrouter.test/v1",
+            developerOverrideEnabled: true
+        )
+
+        model.applySettings(config: config, trustedRouterAPIKeyConfigured: true)
+
+        XCTAssertEqual(model.root.config, config)
+        XCTAssertEqual(model.selectedThread?.mode, .review)
+        XCTAssertEqual(model.selectedThread?.model, "z-ai/glm-5.2")
+        XCTAssertEqual(model.surface().settings.apiBaseURL, "https://api.trustedrouter.test/v1")
+        XCTAssertTrue(model.surface().settings.developerOverrideEnabled)
+        XCTAssertTrue(model.surface().settings.hasStoredAPIKey)
+        XCTAssertEqual(model.surface().settings.apiKeyStatusLabel, "API key configured")
+    }
+
+    func testBootstrapPersistsAndClearsTrustedRouterAPIKey() throws {
+        let paths = QuillCodePaths(home: try makeTempDirectory())
+        let bootstrap = QuillCodeWorkspaceBootstrap(paths: paths)
+
+        XCTAssertFalse(bootstrap.hasTrustedRouterAPIKey())
+        try bootstrap.saveTrustedRouterAPIKey("  sk-tr-v1-test  ")
+        XCTAssertTrue(bootstrap.hasTrustedRouterAPIKey())
+
+        let model = try bootstrap.makeModel()
+        XCTAssertTrue(model.surface().settings.hasStoredAPIKey)
+
+        try bootstrap.clearTrustedRouterAPIKey()
+        XCTAssertFalse(bootstrap.hasTrustedRouterAPIKey())
+    }
+
+    private func makeTempDirectory() throws -> URL {
+        let url = URL(fileURLWithPath: NSTemporaryDirectory())
+            .appendingPathComponent("QuillCodeAppTests-\(UUID().uuidString)")
+        try FileManager.default.createDirectory(at: url, withIntermediateDirectories: true)
+        return url
+    }
+}
