+++ docs/DECISIONS.md
@@ -208,6 +208,7 @@
 - Multi-agent collaboration should happen through pull requests and a serialized merge train rather than direct pushes to `main`. CI runs on `pull_request`, `push`, and `merge_group` events. The train watches `merge-train` and `automerge` labels, ignores draft or `do-not-merge` PRs, processes only the oldest eligible PR, updates behind branches before merging, and merges only after all non-train checks are successful. This keeps Codex/agent work reviewable while preserving a green mainline.
 - The first memories slice is auditable and user-authored. QuillCode loads bounded `.md`, `.txt`, and `.json` notes from `~/.quillcode/memories` and project `.quillcode/memories`, rejects symlinks and paths outside the memory root, truncates oversized notes with visible labels, snapshots loaded memories onto threads, injects them into TrustedRouter as non-command background context, and surfaces the count/sources through the top bar, Memories pane, static HTML renderer, Playwright harness, slash command, and macOS menu bar. Users can explicitly save global memories with `/remember text`, the Add memory command-palette action, or a model-authored `host.memory.remember` tool call when they ask QuillCode to remember a stable preference/fact; saved notes create transcript/tool audit records, refresh memory context immediately, and reject obvious credentials, tokens, passwords, and private keys. Global memories can be forgotten from the Memories pane and produce a transcript notice. Project memories stay read-only in-app because they are project-owned files. Richer redaction review, idle Chronicle jobs, editing UI, conflict resolution, and fully autonomous inference of memories without explicit user intent are separate safety/UX milestones.
 - Workspace configuration transitions live in `WorkspaceConfigurationEngine`, not inline in `WorkspaceModel`. Mode/model selection, model ID normalization, favorite-model mutation, catalog normalization, settings application, and selected-thread config syncing are pure state rules that need direct tests and parity gates. The workspace model owns user-facing orchestration: when to refresh the top bar, persist stores, or route UI commands.
+- Workspace configuration integration coverage lives in `WorkspaceConfigurationIntegrationTests`, not `WorkspaceModelTests`. Mode/model propagation to the selected thread and top bar, favorite model surface updates, settings application, and TrustedRouter key persistence are one settings/configuration feature group; keep them together while pure reducer behavior stays in `WorkspaceConfigurationEngineTests`.
 - Tool audit event construction lives in `WorkspaceToolEventRecorder`, not inline in `WorkspaceModel`. Queued/running/completed/failed tool events, redacted call payloads, and result payload JSON are transcript contracts shared by native tool cards, static HTML, activity, and tests. The workspace model owns only when a tool run should be recorded.
 - Tool-call execution routing lives in `WorkspaceToolCallExecutor`. Browser inspect, plan update, SSH Remote project dispatch, local `ToolRouter` fallback, and successful apply-patch review-diff follow-up share one tested routing boundary. `WorkspaceModel` owns context refresh, event recording, persistence, and top-bar status, not tool-name routing branches.
 - Project metadata loading lives in `WorkspaceProjectMetadataLoader`, not inline in `WorkspaceModel`. Local project refresh aggregates instructions, local environment actions, extension manifests, and project memories through one loader; SSH Remote refresh converts remote context into the same metadata contract with local-only actions/extensions cleared. `WorkspaceModel` owns selection, thread-context application, persistence, and top-bar refresh after metadata is loaded.
